requirements.txt
Flask
Werkzeug
requests_html
Pillow
gunicorn
requests
apscheduler